fix: 기존 upStationId와 downStationId 필드의 데이터 타입을 int에서 Long으로 변경
- fix: StationLine 엔티티 클래스에서 List<StationSection> 타입을 갖는 필드 추가 및 1:N 관계 맵핑

+++ src/main/java/subway/dto/StationLineResponse.java
@@ -4,13 +4,13 @@ public class StationLineResponse {
     private final Long id;
     private final String name;
     private final String color;
-    private final int upStationId;
+    private final Long upStationId;
 
-    private final int downStationId;
+    private final Long downStationId;
 
     private final int distance;
 
-    public StationLineResponse(Long id, String name, String color, int upStationId, int downStationId, int distance) {
+    public StationLineResponse(Long id, String name, String color, Long upStationId, Long downStationId, int distance) {
         this.id = id;
         this.name = name;
         this.color = color;
@@ -31,11 +31,11 @@ public class StationLineResponse {
         return color;
     }
 
-    public int getUpStationId() {
+    public Long getUpStationId() {
         return upStationId;
     }
 
-    public int getDownStationId() {
+    public Long getDownStationId() {
         return downStationId;
     }
